correcion de errores con el compilador y actualizacion del bean de login ademas de los estilos y semantica del login
